packages/commonlibs/pyproject.toml
[project]
name = "commonlibs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Xudong Zhang", email = "zhangxudong@lixiang.com" }
]
requires-python = ">=3.10"
dependencies = [
    "psutil>=7.0.0",
    "ruamel-yaml>=0.18.10",
]

[[tool.uv.index]]
url = "https://artifactory.ep.chehejia.com/artifactory/api/pypi/pypi-remote/simple"
default = true

[tool.pytest.ini_options]
log_cli = "true"
log_cli_level = "DEBUG"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["commonlibs"]
